Scripts follow style guidelines in Abstract/Data
