+++ Tower Defence/Abstract/Data/DuplicateTypes.cs
@@ -1,0 +1,11 @@
+namespace BestagonDefense.Abstract.Data;
+
+/// <summary>
+/// A way to disable duplicates in random selections
+/// </summary>
+public enum DuplicateTypes
+{
+    None,
+    ByName,
+    ByType
+}
