+++ Tower Defence/Abstract/Data/Strain.cs
@@ -1,0 +1,10 @@
+namespace BestagonDefense.Abstract.Data;
+
+/// <summary>
+/// The type/variant/strain of Weighted(Curved)List
+/// </summary>
+public enum Strain
+{
+    TurretBlueprint,
+    ModuleChainHandler
+}
