--- Tower Defence/Abstract/ISubtypeable.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace BestagonDefense.Abstract;
-
-public interface ISubtypeable
-{
-    public Type GetSubtype();
-}
